Rewrote game to be more efficient and reduce memory issues
There are many changes including a new structure of the program.
The main difference is a game struct that houses many of the
GtkWidgets instead of having many globals in the main. Tile and
Board no longer handle visuals in any way. As well, the double
array of Tiles for the Board/grid is now a just a longer array.